Add the ability to see all of the API requests
We do this by extending the Exceptions Controller and copying the views (I didn't have the time to make this "prettier"). We extend the ExceptionsStores to be able to accept a flag to fetch the errors only (since we're on the same table, just different error level). This makes it so that we reuse alll of the Exception code.